Fix some non-portable aspects of tests
