+++ tests/__init__.py
@@ -1,0 +1,34 @@
+from pathlib import Path
+from shutil import rmtree
+
+from mora_the_explorer import Config, Explorer
+
+
+TEST_DIR = Path(__file__).parent
+MOCK_APP_CONFIG = TEST_DIR / "mock_app_config.toml"
+MOCK_USER_CONFIG = TEST_DIR / "mock_user_config.toml"
+MOCK_SERVER = TEST_DIR / "mock_server"
+MOCK_DEST = TEST_DIR / "nmr"
+
+def mock_config() -> Config:
+    """Creates a `Config` from the two mock config files written for testing."""
+    return Config(MOCK_APP_CONFIG, MOCK_USER_CONFIG)
+
+def mock_explorer() -> Explorer:
+    """Creates an `Explorer` using the `Config` generated by `mock_config()`."""
+    MOCK_DEST.mkdir(exist_ok=True)
+    empty_folder(MOCK_DEST)
+    config = mock_config()
+    config.paths.windows = str(MOCK_SERVER)
+    config.paths.darwin = str(MOCK_SERVER)
+    config.paths.linux = str(MOCK_SERVER)
+    config.paths.save = str(MOCK_DEST)
+    explorer = Explorer(config)
+    return explorer
+
+def empty_folder(path: Path):
+    for x in path.iterdir():
+        if x.is_file():
+            x.unlink()
+        elif x.is_dir():
+            rmtree(x)
